[#4403][refactor] Move fusion, kvcache, and compile to modular inference optimizer (#126)
* move fusions


* move rms_norm fusion


* move rms_norm fusion


* move kvcache


* move compile_model


* address review comments


* fix: correct rms_norm backend


---------
